Added Bullet
Added Bullet, spawns at players position and fires at a rate of 8fps.
Uses a prefab (Same instance) to stop the game recreating the image and
deleting it which saves on space.
Also made it so when the bullet goes off screen the bullet destorys to
stop loads of bullets building up.

To do:
Enemies